refactor: invalid role 테스트코드 추가, Theme admin controller 분리
