+++ src/main/java/nextstep/exception/NoSuchRoleException.java
@@ -1,5 +1,9 @@
 package nextstep.exception;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
 public class NoSuchRoleException extends RuntimeException {
     public NoSuchRoleException() {
         super("해당 역할은 존재하지 않습니다.");
