+++ src/main/java/nextstep/theme/controller/ThemeController.java
@@ -16,6 +16,7 @@ import java.net.URI;
 import java.util.List;
 
 @RestController
+@RequestMapping("/themes")
 public class ThemeController {
     private ThemeService themeService;
 
@@ -23,22 +24,9 @@ public class ThemeController {
         this.themeService = themeService;
     }
 
-    @PostMapping("/admin/themes")
-    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
-        Long id = themeService.create(themeRequest);
-        return ResponseEntity.created(URI.create("/themes/" + id)).build();
-    }
-
-    @GetMapping("/themes")
+    @GetMapping
     public ResponseEntity<List<Theme>> showThemes() {
         List<Theme> results = themeService.findAll();
         return ResponseEntity.ok().body(results);
-    }
-
-    @DeleteMapping("/admin/themes/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
-        themeService.delete(id);
-
-        return ResponseEntity.noContent().build();
     }
 }
